+++ tests/test_incremental.py
@@ -13,9 +13,28 @@ import dask_ml.metrics
 from dask_ml.metrics.scorer import check_scoring
 
 
+def test_get_params():
+    clf = Incremental(SGDClassifier())
+    result = clf.get_params()
+
+    assert 'estimator__max_iter' in result
+    assert result['scoring'] is None
+
+
+def test_set_params():
+    clf = Incremental(SGDClassifier())
+    clf.set_params(**{'scoring': 'accuracy',
+                      'estimator__max_iter': 20})
+    result = clf.get_params()
+
+    assert result['estimator__max_iter'] == 20
+    assert result['scoring'] == 'accuracy'
+
+
 def test_incremental_basic(scheduler, xy_classification):
     X, y = xy_classification
-    with scheduler() as (s, [a, b]):
+
+    with scheduler() as (s, [_, _]):
         est1 = SGDClassifier(random_state=0, tol=1e-3)
         est2 = clone(est1)
 
@@ -26,11 +45,12 @@ def test_incremental_basic(scheduler, xy_classification):
 
         assert result is clf
 
-        assert isinstance(result.estimator.coef_, np.ndarray)
-        np.testing.assert_array_almost_equal(result.estimator.coef_,
+        assert isinstance(result.estimator_.coef_, np.ndarray)
+        np.testing.assert_array_almost_equal(result.estimator_.coef_,
                                              est2.coef_)
 
-        assert_estimator_equal(clf.estimator, est2, exclude=['loss_function_'])
+        assert_estimator_equal(clf.estimator_, est2,
+                               exclude=['loss_function_'])
 
         #  Predict
         result = clf.predict(X)
@@ -46,31 +66,18 @@ def test_incremental_basic(scheduler, xy_classification):
 
         clf = Incremental(SGDClassifier(random_state=0, tol=1e-3))
         clf.partial_fit(X, y, classes=[0, 1])
-        assert_estimator_equal(clf.estimator, est2, exclude=['loss_function_'])
+        assert_estimator_equal(clf.estimator_, est2,
+                               exclude=['loss_function_'])
 
 
 def test_in_gridsearch(scheduler, xy_classification):
     X, y = xy_classification
+    clf = Incremental(SGDClassifier(random_state=0, tol=1e-3))
+    param_grid = {'estimator__alpha': [0.1, 10]}
+    gs = sklearn.model_selection.GridSearchCV(clf, param_grid, iid=False)
+
     with scheduler() as (s, [a, b]):
-        clf = Incremental(SGDClassifier(random_state=0, tol=1e-3))
-        param_grid = {'alpha': [0.1, 10]}
-        gs = sklearn.model_selection.GridSearchCV(clf, param_grid, iid=False)
         gs.fit(X, y, classes=[0, 1])
-
-
-def test_estimator_param_raises():
-
-    class Dummy(sklearn.base.BaseEstimator):
-        def __init__(self, estimator=42):
-            self.estimator = estimator
-
-        def fit(self, X):
-            return self
-
-    clf = Incremental(Dummy(estimator=1))
-
-    with pytest.raises(ValueError, match='used by both'):
-        clf.get_params()
 
 
 def test_scoring(scheduler, xy_classification,
@@ -96,22 +103,20 @@ def test_scoring_string(scheduler, xy_classification, scoring):
         assert callable(check_scoring(clf, scoring=scoring))
         clf.fit(X, y, classes=np.unique(y))
         clf.score(X, y)
-        clf.estimator.score(X, y)
 
 
 def test_fit_ndarrays():
     X = np.ones((10, 5))
-    y = np.ones(10)
+    y = np.concatenate([np.zeros(5), np.ones(5)])
 
     sgd = SGDClassifier(tol=1e-3)
     inc = Incremental(sgd)
 
     inc.partial_fit(X, y, classes=[0, 1])
-    inc.fit(X, y)
+    sgd.fit(X, y)
 
     assert inc.estimator is sgd
-    assert (sgd.predict(X) == y).all()
-    assert_eq(inc.coef_, inc.estimator.coef_)
+    assert_eq(inc.coef_, inc.estimator_.coef_)
 
 
 def test_score_ndarrays():
@@ -122,7 +127,7 @@ def test_score_ndarrays():
     inc = Incremental(sgd, scoring='accuracy')
 
     inc.partial_fit(X, y, classes=[0, 1])
-    inc.fit(X, y)
+    inc.fit(X, y, classes=[0, 1])
 
     assert inc.score(X, y) == 1