+++ docs/source/changelog.rst
@@ -15,6 +15,21 @@ API Breaking Changes
 --------------------
 
 - Removed the ``basis_inds_`` attribute from :class:`dask_ml.cluster.SpectralClustering` as its no longer used (:pr:`152`)
+- Change :meth:`dask_ml.wrappers.Incremental.fit` to clone the underlying estimator before training (:pr:`258`). This induces a few changes
+
+  1. The underlying estimator no longer gives access to learned attributes like ``coef_``. We recommend using
+     ``Incremental.coef_``.
+  2. State no longer leaks between successive ``fit`` calls. Note that :meth:`Incremental.partial_fit` is still available
+     if you want state, like learned attributes or random seeds, to be re-used. This is useful if you're making multiple
+     passes over the training data.
+- Changed ``get_params`` and ``set_params`` for :class:`dask_ml.wrappers.Incremental` to no longer magically get / set parameters
+  for the underlying estimator (:pr:`258`). To specify parameters for the underlying estimator, use the double-underscore prefix convention
+  established by scikit-learn:
+
+  .. code-block:: python
+
+     inc.set_params('estimator__alpha': 10)
+
 
 Version 0.6.0
 ~~~~~~~~~~~~~